+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 ### v1.7.24 / 2019-08-29
 
+* [114](https://github.com/b3log/vditor/issues/114) 封装 Preview `enhancement`
 * [111](https://github.com/b3log/vditor/issues/111) emoji problem `bug`
 * [110](https://github.com/b3log/vditor/issues/110) 支持 Word 粘贴 `feature`
 * [109](https://github.com/b3log/vditor/issues/109) 移除 option.editorName 变量 `bug`